requirements.txt
requests
beautifulsoup4
peewee
# PyMySQL
psycopg2
flask
honcho
python-dotenv
